move `is_moderator()` to `views.py`, and add checks to moderator pages
